requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.6

-e .